Rename binance_scr.ipynb to crypto_trader_helper.ipynb
It helps crypto traders to save the prices at the moment they buy or convert their assets (Trade class). And shows which conversion is beneficial by using second class - Profit. If select_only set to true it will show only profitable conversions. If select set to true it will want a coin and return data for that specific token/coin. If return_ set to true it will return the whole table.